requirements.txt
Django
pillow
psycopg2-binary
django-environ
django-crispy-forms
crispy-bootstrap5
